+++ src/commands/runMacro.ts
@@ -1,4 +1,6 @@
 import * as vscode from 'vscode';
+import { MacroRunner } from '../core/execution/macroRunner';
+import { Macro } from '../core/macro';
 import { ExtensionContext } from '../extensionContext';
 import { showMacroErrorDialog, showMacroQuickPick } from '../ui/dialogs';
 import { activeMacroEditor } from '../utils/activeMacroEditor';
@@ -6,22 +8,27 @@ import { PathLike, toUri } from '../utils/uri';
 
 export async function runMacro(
   { libraryManager, log, runnerManager, mruMacro }: ExtensionContext,
-  pathOrUri?: PathLike,
+  pathOrUriOrMacro?: PathLike | Macro,
   startup?: true,
 ) {
-  const uri = pathOrUri
-    ? toUri(pathOrUri)
-    : await showMacroQuickPick(libraryManager, { selectUri: mruMacro });
-  if (!uri) {
-    return; // Nothing to run.
+  let runner: MacroRunner;
+  if (pathOrUriOrMacro instanceof Macro) {
+    runner = runnerManager.getRunner(pathOrUriOrMacro);
+  } else {
+    const uri = pathOrUriOrMacro
+      ? toUri(pathOrUriOrMacro)
+      : await showMacroQuickPick(libraryManager, { selectUri: mruMacro });
+    if (!uri) {
+      return; // Nothing to run.
+    }
+    runner = runnerManager.getRunner(uri);
   }
 
-  const runner = runnerManager.getRunner(uri);
   const [code, options] = await runner.getMacroCode();
   if (code.length === 0) {
     log.warn(
       'Skipped running macro file: file is empty.',
-      `"${vscode.workspace.asRelativePath(uri, true)}"`,
+      `"${vscode.workspace.asRelativePath(runner.macro.uri, true)}"`,
     );
     return;
   }